gradient-agents/requirements.txt
gradient-adk>=0.1.9
gradient>=3.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
